[feat] Boss3 Platform 공격 패턴 기초작업
- Boss3 의 기본 발판 공격 패턴의 기초 작업
