+++ Assets/Programing/Ji/Boss3Sample.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 464fc07e9cf92134aae59e81ae918575
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
