+++ Assets/Programing/Ji/LenPlatformLoop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32d0dc95e7b26674d9f3dd6c9f96850d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
