+++ Assets/Programing/Ji/MoveLenPlatform.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MoveLenPlatform : MonoBehaviour
+{
+    // 일정한 속도로
+    [SerializeField] float moveSpeed;
+    // 방향별로 이동
+
+
+
+}
