+++ Assets/Programing/Ji/PlatformAttack.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+/// <summary>
+/// 몬스터의 공격 패턴 중 하나로
+/// 랜덤으로 줄 발판들을 위 또는 아래로 움직이도록 설정
+/// </summary>
+public class PlatformAttack : MonoBehaviour
+{
+    
+}
