+++ Assets/Programing/Ji/StartPlatform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b87c5345f8f4ea4eae511901d0403e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
